♻️ MCY8ZC supervisor: keep protocol tests within hotspot budgets
