+++ packages/agentplane/src/commands/shared/route-decision-blockers.ts
@@ -11,7 +11,7 @@ import { isTaskSetLocalOnlyAdvance } from "./task-local-freshness.js";
 import type { CommandContext } from "./task-backend.js";
 import { isRecord } from "../../shared/guards.js";
 import { getHumanInputState } from "../task/human-input.js";
-import { resolveTaskDependencyState } from "../task/shared/dependencies.js";
+import { taskDependencyReadinessBlocker } from "../task/shared/dependencies.js";
 import { parsePrMeta, type PrMeta } from "./pr-meta.js";
 import { readTaskPrArtifact } from "../pr/internal/pr-paths.js";
 import { hasAcceptedQualityReviewProvenance } from "../task/quality-review-gate.js";
@@ -326,20 +326,8 @@ export async function deriveBlockers(opts: {
   }
   const normalizedTaskStatus = String(opts.task.status).toUpperCase();
   if (normalizedTaskStatus === "TODO") {
-    const dependencies = await resolveTaskDependencyState(opts.task, opts.ctx.taskBackend);
-    if (dependencies.missing.length > 0 || dependencies.incomplete.length > 0) {
-      const details = [
-        ...(dependencies.missing.length > 0 ? [`missing: ${dependencies.missing.join(", ")}`] : []),
-        ...(dependencies.incomplete.length > 0
-          ? [`incomplete: ${dependencies.incomplete.join(", ")}`]
-          : []),
-      ];
-      addBlocker(
-        blockers,
-        "dependency_not_ready",
-        `task dependencies are not ready (${details.join("; ")})`,
-      );
-    }
+    const summary = await taskDependencyReadinessBlocker(opts.task, opts.ctx.taskBackend);
+    if (summary) addBlocker(blockers, "dependency_not_ready", summary);
   }
   if (
     opts.workflowMode === "branch_pr" &&
